Fix incorrect padding on posts with lists
We currently merge together all list text blocks (assuming they are
apart of the same list) into one list grouping object. This means that the
total amount of blocks is reduced compared with the raw JSON data hence the
need for padding in order to successfully locate the block from the layout
indices Tumblr gives us.

Prior to this commit, for each list grouping the amount that was padded onto
the total amount of blocks was equal to the amount of items within the list
grouping. However, this behavior is correct.

Take the following NPF data:

text - text
text - list
text - list
text - list
text - text

When it gets parsed it gets turned into the following representation:

text - text
ListGrouping(
    text - list
    text - list
    text - list
    text - list
)
text - text

So by natively padding 4 items to the amount of blocks we end up with:

text - text
ListGrouping
pad
pad
pad
pad
text - text

Which exceeds the total amount of actual blocks by one due to the
additonal ListGrouping block.

The correct behavior would be to only pad 3 times
or the total amount of children - 1

text - text
ListGrouping
pad
pad
pad
text - text

Ending up with a combined total of 6 blocks which is equal to the amount
of actual blocks