Rewrite logic for inline formatting (#5)
This commit simplifies the convoluted method in which inline
formatting was handled.

By splitting the overlapping inline formatting intervals into
discrete slices that signifies how to style that slice, the logic for
inline formatting can be simplified to a single loop instead of the
recursive series of checks and dozens of while loops.

This improves performance, code clarity, and fixes some long-standing
bugs in the renderer.

The caveat however is that the resulting HTML now has the chance to
be much larger than before. This should be addressed as soon as
possible.